more refactoring, getting ready for subcommands
