+++ src/main.rs
@@ -12,18 +12,16 @@ use std::env;
 use std::path;
 use std::ffi::OsString;
 
+use anyhow::{Context, Result};
 use colored::*;
 use rayon::prelude::*;
 
 mod die;
-use die::die;
-
 mod runit;
-//use runit::{RunitService, RunitServiceState};
-
 mod utils;
-
 mod service;
+
+use die::die;
 use service::Service;
 
 const SERVICE_DIR: &str = "/var/service";
@@ -40,26 +38,20 @@ macro_rules! verbose {
 }
 */
 
-fn print_service(service: &Service) {
-    println!("{}", service);
-}
-
-fn main() {
-    colored::control::set_override(COLORIZE);
-    colored::control::unset_override();
-
+fn do_status() -> Result<()> {
+    // get SVDIR from env or use default
     let svdir = env::var_os("SVDIR")
         .unwrap_or_else(|| OsString::from(SERVICE_DIR) );
     let svdir = path::Path::new(&svdir);
 
+    // check if user wants pstree
     let want_pstree = env::var_os("PSTREE").is_some();
 
     // find all services
-    let services = runit::get_services(svdir)
-        .unwrap_or_else(|err| die!(1, "failed to list services: {}", err));
+    let services = runit::get_services(svdir).context("failed to list services")?;
 
     // process each service found (just gather data here, can be done in parallel)
-    let objects: Vec<Service> = services
+    let services: Vec<Service> = services
         .par_iter()
         .filter_map(|service| {
             match Service::from_runit_service(service, want_pstree) {
@@ -74,8 +66,8 @@ fn main() {
 
     // print gathared data
     println!();
-    println!("found {} services in {:?}", objects.len(), svdir); // verbose
-    println!("  {:1} {:15} {:10} {:10} {:10} {:15} {:10}",
+    println!("found {} services in {:?}", services.len(), svdir); // verbose
+    println!("  {:1} {:20} {:7} {:9} {:8} {:17} {}",
         "",
         "SERVICE".bold(),
         "STATE".bold(),
@@ -85,9 +77,24 @@ fn main() {
         "TIME".bold());
 
     // print each service found
-    for object in objects {
-        print_service(&object);
+    for service in services {
+        println!("{}", service);
     }
 
     println!();
+
+    Ok(())
+}
+
+fn main() {
+    // color output
+    colored::control::set_override(COLORIZE);
+    colored::control::unset_override();
+
+    // figure out subcommand to run
+    let ret = do_status();
+
+    if let Err(err) = ret {
+        die!(1, "{}: {:?}", "error".red(), err);
+    }
 }
